Send out messages unprompted
I did test this but the code is commented out (app.py:52). Works great,
though be aware that the numbers you text need to be in your Twilio
sandbox.

I added new environment variables (see README.md, it has them in the
description of launch.json) so that we can all use our own sandboxes
with the push functionality.

Chatbot logic is inside the new class. The two basic behaviors are
replying (send a response to a text) and pushing (send a text to a
number when that number has not texted first to provoke the text). These
are in functions. I was thinking we could use them with bot slash
commands, so I wrote a static method for the class that would be able to
call them. Based on the user's slash command, it would have side effects
like modifying the JSON file or sending out group messages and then it
would return whatever we want to say to the user, e.g., your message was
sent successfully, you have subscribed, here is the translation in x
language of the last message you tried to send, etc. Based on this idea
I have some ideas for what would be handy in the JSON structure; see
Taiga task description.