JSON structure idea + refactors
JSON consists of keys that are WhatsApp contact info ("whatsapp:+<phone
number>") and values that are the profile name associated with the
contact info, that user's preferred language, and their last drafted
message, which should be null if they have not drafted a message or if
their draft was sent.

Common-sense refactors as well as consideration for modifying a JSON
file. It would be easier to work with a dictionary so we don't need to
worry about multiple user requests trying to modify the same file at
once, but we will need a single thread to periodically save changes to
the JSON. Is there a better way to do this? As it stands that leaves a
possibility of the file being not up-to-date if the server shuts down
unexpectedly.