Begin Poll class
I need to figure out how async timers should work and modify the Chatbot
constructor to function with them. I also need to implement the bulk of
the logic to actually process the /poll command and do appropriate error
reporting.

There will have to be some way to vote and unfortunately Twilio doesn't
seem to be capable of detecting replies to messages in WhatsApp, so I
guess it'll have to be via some sort of argument, which is not ideal.